+++ CHANGES.rst
@@ -20,6 +20,7 @@ Development
   Thanks Martin for reporting this glitch and thanks George for conceiving this fine Python package.
 - [ui] Add link to New Espacenet, see #4. Thanks, Felix!
 - [ui] Attempt to mitigate "TypeError: t is null" notifications from production
+- [ui] Fix crawler dialog
 
 
 .. _python-epo-ops-client: https://pypi.org/project/python-epo-ops-client/